functions/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o functions/requirements.txt
anyio==4.12.1
    # via starlette
blinker==1.9.0
    # via flask
certifi==2026.1.4
    # via requests
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via
    #   flask
    #   functions-framework
    #   uvicorn
cloudevents==1.12.0
    # via functions-framework
deprecation==2.1.0
    # via cloudevents
duckdb==1.4.3
    # via cityjobs (pyproject.toml)
flask==3.1.2
    # via functions-framework
functions-framework==3.10.0
    # via cityjobs (pyproject.toml)
google-api-core==2.29.0
    # via
    #   google-cloud-core
    #   google-cloud-secret-manager
    #   google-cloud-storage
google-auth==2.47.0
    # via
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-secret-manager
    #   google-cloud-storage
google-cloud-core==2.5.0
    # via google-cloud-storage
google-cloud-secret-manager==2.26.0
    # via cityjobs (pyproject.toml)
google-cloud-storage==3.7.0
    # via cityjobs (pyproject.toml)
google-crc32c==1.8.0
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.8.0
    # via google-cloud-storage
googleapis-common-protos==1.72.0
    # via
    #   google-api-core
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.14.3
    # via google-cloud-secret-manager
grpcio==1.76.0
    # via
    #   google-api-core
    #   google-cloud-secret-manager
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.76.0
    # via google-api-core
gunicorn==23.0.0
    # via
    #   functions-framework
    #   uvicorn-worker
h11==0.16.0
    # via uvicorn
idna==3.11
    # via
    #   anyio
    #   requests
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
markupsafe==3.0.3
    # via
    #   flask
    #   jinja2
    #   werkzeug
mashumaro==3.17
    # via cityjobs (pyproject.toml)
numpy==2.4.1
    # via pandas
packaging==25.0
    # via
    #   deprecation
    #   gunicorn
pandas==2.3.3
    # via cityjobs (pyproject.toml)
proto-plus==1.27.0
    # via
    #   google-api-core
    #   google-cloud-secret-manager
protobuf==6.33.3
    # via
    #   google-api-core
    #   google-cloud-secret-manager
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
python-dateutil==2.9.0.post0
    # via pandas
pytz==2025.2
    # via pandas
requests==2.32.5
    # via
    #   cityjobs (pyproject.toml)
    #   google-api-core
    #   google-cloud-storage
rsa==4.9.1
    # via google-auth
six==1.17.0
    # via python-dateutil
starlette==0.51.0
    # via functions-framework
typing-extensions==4.15.0
    # via
    #   anyio
    #   grpcio
    #   mashumaro
    #   starlette
tzdata==2025.3
    # via pandas
urllib3==2.6.3
    # via requests
uvicorn==0.40.0
    # via
    #   functions-framework
    #   uvicorn-worker
uvicorn-worker==0.4.0
    # via functions-framework
watchdog==6.0.0
    # via functions-framework
werkzeug==3.1.5
    # via
    #   flask
    #   functions-framework
